requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
python-telegram-bot==13.7
psycopg2-binary==2.9.3